Refactor v2 code
