+++ app/services/v2/rag/constant.py
@@ -1,0 +1,12 @@
+# Constants
+CACHE_TTL = 3600  # 1 hour
+DEFAULT_COMPANY = "Our Company"
+DEFAULT_ATTITUDE = "professional"
+DEFAULT_START_MSG = "Chào bạn, đây là ban tư vấn khách hàng của công ty chúng tôi"
+DEFAULT_END_MSG = "Cảm ơn quý khách đã trò chuyện. Hẹn gặp lại quý khách trong thời gian sớm nhất!"
+
+# Cache configurations
+RETRIEVER_CACHE_SIZE = 100
+HISTORY_RETRIEVER_CACHE_SIZE = 100
+PROMPT_CACHE_SIZE = 100
+QA_CACHE_SIZE = 100
