+++ app/utils/lifespan.py
@@ -1,7 +1,5 @@
 from contextlib import asynccontextmanager
-
 from fastapi import FastAPI
-
 from app.core.db import mongo_connection
 
 @asynccontextmanager
