--- app/utils/metadata.json
@@ -1,16 +1,0 @@
-{
-    "files": {
-        "Policy.docx": 3, 
-        "ReviewCV-policy.pdf": 3
-    },
-    "urls": [
-        {
-            "url": "https://itviec.com/blog/faq-chuc-nang-loi-moi-cong-viec-danh-cho-ung-vien/", 
-            "priority": 2
-        },
-        {
-            "url": "https://itviec.com/blog/faq-ve-chuc-nang-company-review/", 
-            "priority": 2
-        }
-    ]
-}
