+++ pyproject.toml
@@ -18,6 +18,7 @@ dependencies = [
     "langchain>=0.3.13",
     "langchain-community>=0.3.21",
     "langchain-core>=0.3.28",
+    "langchain-google-genai>=2.1.4",
     "langchain-openai>=0.2.14",
     "langchain-pinecone>=0.2.0",
     "langgraph>=0.2.74",
