--- tests/llm_test/gen_api.py
@@ -1,5 +1,0 @@
-import secrets
-
-api_key = secrets.token_urlsafe(45)
-print("coolchat-proj-"+api_key)
-# cool-proj-N-vGdWSO6tAXOTIthcV9vRVmEti3HkNe5NOU8QiFXBwJcdave6Twin6-9XzE
